Add basic error handling
A function can be marked as fallible by putting a ? after the function name:

`fn test?(x int) {}`

Then use the `fail` keyword to fail from such a function.

Currently the only way to handle errors is by propagating the error like this:

`test(10)?`

If this is called in the main or another infalliblefunction, the program will
just exit with the error message.

Actually recovering from errors will be implemeted later.